requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
greenlet==3.2.4
holidays==0.59
idna==3.11
iniconfig==2.1.0
packaging==25.0
playwright==1.55.0
playwright-stealth==2.0.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
